pyproject.toml
[project]
name = "api-ui-pytest-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Elena",email = "84013140+ElenaShorohova@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "selenium (>=4.29.0,<5.0.0)",
    "parse (>=1.20.2,<2.0.0)",
    "jsonpath-rw-ext (>=1.2.2,<2.0.0)"
]

[tool.poetry]
package-mode = false
name = "api-ui-pytest-demo"
version = "0.1.0"
description = ""
authors = ["Шорохова Елена <eshorohova89@gmail.com>"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
allure-pytest = "^2.13.5"
flake8 = "^7.1.2"
black = "^25.1.0"
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.poetry.dependencies]
python = ">=3.12,<4.0"

[tool.flake8]
max-line-length = 120